requirements.txt
argparse
flowmetricscsv
montecarlocsv
azure-devops
requests
wheel
twine
setuptools
pytest
pytest-cov
flake8
black